pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mqi_interface"
version = "0.1.0"
authors = [
  { name="Your Name", email="you@example.com" },
]
description = "A system for managing and orchestrating medical physics QA workflows."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pika==1.3.1",
    "paramiko==3.4.0",
    "PyYAML==6.0.1",
    "APScheduler==3.10.1",
    "psutil==5.9.5",
    "cryptography==43.0.1",
    "structlog==23.1.0",
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "jinja2==3.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.0",
    "pytest-asyncio==0.21.0",
    "pytest-cov==4.1.0",
    "black==23.3.0",
    "ruff==0.1.9",
    "mypy==1.4.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests/unit",
    "tests/integration",
]
pythonpath = "src"
addopts = "-ra -q"

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
show_missing = true
